Fix backward compatibility
